+++ app/src/main/java/woowacourse/movie/activity/MainActivity.kt
@@ -1,8 +1,11 @@
-package woowacourse.movie
+package woowacourse.movie.activity
 
 import android.os.Bundle
 import android.widget.ListView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.MovieMockData
+import woowacourse.movie.R
+import woowacourse.movie.adapter.MovieListAdapter
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
