+++ app/src/main/java/woowacourse/movie/adapter/MovieListAdapter.kt
@@ -1,10 +1,13 @@
-package woowacourse.movie
+package woowacourse.movie.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import woowacourse.movie.BundleKeys.MOVIE_DATA_KEY
+import woowacourse.movie.Movie
+import woowacourse.movie.R
+import woowacourse.movie.activity.MovieDetailActivity
 
 class MovieListAdapter(private val movies: List<Movie>) : BaseAdapter() {
     override fun getCount(): Int = movies.size
