+++ app/src/main/res/layout/activity_book_complete.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BookCompleteActivity">
+    tools:context=".activity.BookCompleteActivity">
 
     <TextView
         android:id="@+id/tv_book_complete"
